Move ClientActionStateResult to SimpleSchedulerStateManager
To help with pre-work needed for redis scheduler we make this struct
usable in mods outside of MemoryAwaitedActionDb.

towards #359
